# app/docs/swagger.yaml
basePath: /api
definitions:
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      phone_number:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления GoGRAFF
  title: GoGRAFF API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Возвращает список всех пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить всех пользователей
      tags:
      - Пользователи
    post:
      consumes:
      - application/json
      description: Создает нового пользователя
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать пользователя
      tags:
      - Пользователи
  /users/{id}:
    delete:
      description: Удаляет пользователя
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Пользователь успешно удален
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Удалить пользователя
      tags:
      - Пользователи
    get:
      description: Возвращает пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить пользователя
      tags:
      - Пользователи
    put:
      consumes:
      - application/json
      description: Обновляет данные пользователя
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно обновлен
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновить пользователя
      tags:
      - Пользователи
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"